Add files via upload
In this task,I have created a landing web
page using basic HTML and CSS. This is a
very basic project for beginners. After
completion of this project, you will be able
to level up your CSS. In this project, I have
learnt how to divide a page into the header,
footer, body part, and griding system in
web page.